Move TTY wrapper to subprocess package, makes more sense now it exists
