+++ subprocess/pty.go
@@ -1,4 +1,4 @@
-package git
+package subprocess
 
 import (
 	"io"
