+++ subprocess/pty_windows.go
@@ -1,4 +1,4 @@
-package git
+package subprocess
 
 import "os/exec"
